Add throughput and duration graphs to ActiveJob
The Sidekiq dashboard also has these graphs, I copied them from there.
It shows the throughput for all actions in the "background" namespace
and the duration for those actions in the "background" namespace.

These two graphs make the dashboard more complete and provides more
information based on the `transaction_duration` metric.

Also be more clear that the priority graph can be empty if the adapter
doesn't support it.