Add run queue length to Erlang magic dashboard
The run queue length for a given scheduler is an interesting metric
to be able to observe, especially alongside the scheduler
utilization metric. The metric is already being reported by the
Erlang probe in the Elixir integration, but it is not displayed in
the magic dashboard. This commit adds a graph of run queue lengths
to the magic dashboard.